+++ Sources/BrazeUI/InAppMessageUI/Views/InAppMessageUIModalView.swift
@@ -51,9 +51,13 @@ extension BrazeInAppMessageUI {
       public var minWidth = 320.0
 
       /// The maximum width.
+      ///
+      /// The maximum width is swapped with the maximum height for large UIs (e.g. iPad).
       public var maxWidth = 450.0
 
       /// The maximum height.
+      ///
+      /// The maximum height is swapped with the maximum width for large UIs (e.g. iPad).
       public var maxHeight = 720.0
 
       /// Specify whether the modal in-app message view can be dismissed from a tap with the
@@ -128,9 +132,11 @@ extension BrazeInAppMessageUI {
       shadowView.shadow = attributes.shadow
 
       // Dimensions
+      let maxWidth = isLargeLandscape ? attributes.maxHeight : attributes.maxWidth
+      let maxHeight = isLargeLandscape ? attributes.maxWidth : attributes.maxHeight
       minWidthConstraint.constant = attributes.minWidth
-      maxWidthConstraint.constant = attributes.maxWidth
-      maxHeightConstraint.constant = attributes.maxHeight
+      maxWidthConstraint.constant = maxWidth
+      maxHeightConstraint.constant = maxHeight
 
       // User interactions
       tapBackgroundGesture.isEnabled = attributes.dismissOnBackgroundTap
@@ -229,6 +235,7 @@ extension BrazeInAppMessageUI {
       view.stack.axis = .vertical
       view.stack.distribution = .fill
       view.stack.isLayoutMarginsRelativeArrangement = true
+      view.layer.masksToBounds = true
       view.addSubview(closeButton)
       return view
     }()
@@ -296,6 +303,7 @@ extension BrazeInAppMessageUI {
       _ previousTraitCollection: UITraitCollection?
     ) {
       super.traitCollectionDidChange(previousTraitCollection)
+      applyAttributes()
       applyTheme()
     }